refactor: reataurant, trail 컴포넌트 파일 폴더링 및 type 별도 분리
- component 안의 rstaurant와 trail 관련 컴포넌트들을 별도 폴더로 분리하여 관리 체계화
- 컴포넌트 및 타입 분리하여 유지보수성 및 가독성 향상
