+++ src/app/(header-layout)/place/components/trail/TrailContents.tsx
